fix(list-of-companes-component)Check if company has generated item entry. If entry is generated then can't remove that company. Remove entry of thah company first then remove company

+++ src/app/companies/list-of-companies/list-of-companies.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
-import { Observable, tap } from 'rxjs';
+import { map, Observable, tap } from 'rxjs';
 import { CompaniesService } from 'src/app/core/services/companies.service';
 import { Company } from 'src/app/models/company';
 import { CompanyDetailsComponent } from '../company-details/company-details.component';
@@ -46,9 +46,24 @@ export class ListOfCompaniesComponent {
     this.dialog.open(CompanyDetailsComponent, { data: companyDetails });
   }
 
-  //Remove company from databse
-  removeItem(key: string) {
-    this.comapniesService.removeCompany(key).then(this.onRemoveSuccess.bind(this), this.onError.bind(this));
+  //Remove company from databse - check, if can delete (if company has add item entry, then must remove that entry first before remove company)
+  removeItem(key: string, id) {
+    const id$ = this.comapniesService.getItemsInfos().pipe(map(res => res.map(res => res.company_id)));
+    id$.subscribe(res => {
+      if (res.find(el => el == id)) {
+        console.log('Nie można usunąć');
+        this.toast.open(
+          'Nie można usunąć tej firmy ponieważ ma już wygenerowane zestawienia. Najpierw usuń zestawienia dotyczące tej firmy!',
+          '',
+          {
+            panelClass: 'toast-error'
+          }
+        );
+      } else {
+        console.log('Można usunąć');
+        this.comapniesService.removeCompany(key).then(this.onRemoveSuccess.bind(this), this.onError.bind(this));
+      }
+    });
   }
 
   //Message that  item data has property deleted from a server and then router to /dashboard
